backport of commit 47c77a495087a4e37162a6560b6331be94f7266e
