+++ control-plane/api-gateway/common/secrets.go
@@ -12,7 +12,11 @@ import (
 
 	"github.com/miekg/dns"
 	corev1 "k8s.io/api/core/v1"
+
+	"github.com/hashicorp/consul-k8s/control-plane/version"
 )
+
+var errFailedToParsePrivateKeyPem = errors.New("failed to parse private key PEM")
 
 func ParseCertificateData(secret corev1.Secret) (cert string, privateKey string, err error) {
 	decodedPrivateKey := secret.Data[corev1.TLSPrivateKeyKey]
@@ -20,7 +24,7 @@ func ParseCertificateData(secret corev1.Secret) (cert string, privateKey string,
 
 	privateKeyBlock, _ := pem.Decode(decodedPrivateKey)
 	if privateKeyBlock == nil {
-		return "", "", errors.New("failed to parse private key PEM")
+		return "", "", errFailedToParsePrivateKeyPem
 	}
 
 	certificateBlock, _ := pem.Decode(decodedCertificate)
@@ -64,5 +68,52 @@ func validateCertificateHosts(certificate *x509.Certificate) error {
 		}
 	}
 
+	return nil
+}
+
+// Envoy will silently reject any keys that are less than 2048 bytes long
+// https://github.com/envoyproxy/envoy/blob/main/source/extensions/transport_sockets/tls/context_impl.cc#L238
+const MinKeyLength = 2048
+
+// ValidateKeyLength ensures that the key length for a certificate is of a valid length
+// for envoy dependent on if consul is running in FIPS mode or not
+func ValidateKeyLength(privateKey string) error {
+	privateKeyBlock, _ := pem.Decode([]byte(privateKey))
+
+	if privateKeyBlock == nil {
+		return errFailedToParsePrivateKeyPem
+	}
+
+	if privateKeyBlock.Type != "RSA PRIVATE KEY" {
+		return nil
+	}
+
+	key, err := x509.ParsePKCS1PrivateKey(privateKeyBlock.Bytes)
+	if err != nil {
+		return err
+	}
+
+	keyBitLen := key.N.BitLen()
+
+	if version.IsFIPS() {
+		return fipsLenCheck(keyBitLen)
+	}
+
+	return nonFipsLenCheck(keyBitLen)
+}
+
+func nonFipsLenCheck(keyLen int) error {
+	// ensure private key is of the correct length
+	if keyLen < MinKeyLength {
+		return errors.New("RSA key length must be at least 2048-bit")
+	}
+
+	return nil
+}
+
+func fipsLenCheck(keyLen int) error {
+	if keyLen != 2048 && keyLen != 3072 && keyLen != 4096 {
+		return errors.New("RSA key length must be at either 2048-bit, 3072-bit, or 4096-bit in FIPS mode")
+	}
 	return nil
 }
